nvim/plugins/bootstrap: install package manager if fresh system
This module bootstraps our plugin manager

It can (and should) be called before any plugins are initialized,
containing only a dependency on system git, for cloning the package
manager locally.

The 'install_path' variable was taken from packer.nvim's documentation
and installs packer as an optional (opt) plugin. This is important for
when we have packer manage itself, as we'll need to:

1. Ensure it's loaded during bootstrap via `packadd packer`
2. Install load hooks on it's various commands + module