vimrc: fix coc.nvim bindings for >= v0.0.80
coc.nvim introduced a metric ton of backwards incompatible changes to
various APIs they expose in Vim.

This commit updates the bits we use to use the new way of doing things.

It also adds a :Format command that can be used to format the current
buffer.

You'll need to do a :PlugUpdate if your psoxizsh has an old coc locally,
you can check your version with :CocInfo, if the `version` is <= v0.0.79