nvim/plug: +fidget@master
This is a plugin for reporting '$/progress' status messages from neovim.

Realistically, only (some) LSP servers actually use this, but it
provides a clean, minimal method for understanding what an LSP is
working on at the minute.

`Server.Rust`, `Server.Lua` and `Server.Json` support it at a minimum,
from my personal testing