nvim: remove vimpeccable (plug/vimp)
This commit removes all vimpeccable (vimp) usage from the code base, as
it has been superseded by 'psoxizsh.key'.

In addition to simplifying the code, this also removes the annoying
spurious errors that would be generated when reloading configuration.

One thing to note: this commit also remove most of the LSP config we
have for coc.nvim. This is intentional, and will be corrected via a
future patchset utilizing the native neovim LSP client, as coc has
become annoying to maintain.